+++ pom.xml
@@ -9,10 +9,10 @@
 		<relativePath/>
 	</parent>
 	<groupId>de.thowl</groupId>
-	<artifactId>notetilus</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>notetilus</name>
-	<description>Demo project for Spring Boot</description>
+	<artifactId>TnT</artifactId>
+	<version>0.0.1</version>
+	<name>TnT</name>
+	<description>Todo's 'n' Texts</description>
 	<properties>
 		<java.version>21</java.version>
 	</properties>
