--- src/main/java/de/thowl/notetilus/NotetilusApplication.java
@@ -1,15 +1,0 @@
-package de.thowl.notetilus;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.actuate.autoconfigure.security.servlet.ManagementWebSecurityAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-
-@SpringBootApplication(exclude = { SecurityAutoConfiguration.class, ManagementWebSecurityAutoConfiguration.class })
-public class NotetilusApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NotetilusApplication.class, args);
-	}
-
-}
