--- src/main/java/de/thowl/notetilus/core/exeptions/InvalidCredentialsException.java
@@ -1,8 +1,0 @@
-package de.thowl.notetilus.core.exeptions;
-
-import lombok.experimental.StandardException;
-
-@StandardException
-public class InvalidCredentialsException extends Exception {
-
-}
