--- src/main/java/de/thowl/notetilus/storage/GroupRepository.java
@@ -1,12 +1,0 @@
-package de.thowl.notetilus.storage;
-
-import org.springframework.stereotype.Repository;
-import org.springframework.data.repository.CrudRepository;
-import de.thowl.notetilus.storage.entities.Group;
-
-@Repository
-public interface GroupRepository extends CrudRepository<Group, Long> {
-	public Group findById(int id);
-
-	public Group findByName(String name);
-}
