--- src/main/java/de/thowl/notetilus/storage/SessionRepository.java
@@ -1,13 +1,0 @@
-package de.thowl.notetilus.storage;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import de.thowl.notetilus.storage.entities.Session;
-
-@Repository
-public interface SessionRepository extends CrudRepository<Session, Long> {
-	public Session findById(long id);
-	public Session findByAuthToken(String authToken);
-	public Session findByUserId(long userId);
-}
