--- src/main/java/de/thowl/notetilus/storage/entities/AccessToken.java
@@ -1,12 +1,0 @@
-package de.thowl.notetilus.storage.entities;
-
-import java.util.Date;
-
-import lombok.Data;
-
-@Data
-public class AccessToken {
-	private String usid; // unique session id
-	private Date lastActive;
-	private long userId;
-}
