--- src/main/java/de/thowl/notetilus/storage/entities/Group.java
@@ -1,31 +1,0 @@
-package de.thowl.notetilus.storage.entities;
-
-import java.util.List;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
-
-@Data
-@Entity
-@Table(name = "Groups")
-public class Group {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
-
-	@OneToMany(mappedBy = "group")
-	public List<User> members;
-
-	@NotNull
-	private String name;
-
-	private boolean admin;
-
-}
