--- src/main/java/de/thowl/notetilus/storage/entities/User.java
@@ -1,50 +1,0 @@
-package de.thowl.notetilus.storage.entities;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.validation.constraints.NotNull;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@Entity
-@Table(name = "Users")
-@NoArgsConstructor
-@RequiredArgsConstructor
-public class User {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
-
-	@ManyToOne
-	@JoinColumn(name = "group_id", nullable = false)
-	private Group group;
-
-	@NotNull
-	@NonNull
-	private String firstname;
-
-	@NotNull
-	@NonNull
-	private String lastname;
-
-	@NotNull
-	@NonNull
-	private String username;
-
-	@NotNull
-	@NonNull
-	private String email;
-
-	@NotNull
-	@NonNull
-	private String password;
-}
