--- src/main/java/de/thowl/notetilus/web/ForbiddenException.java
@@ -1,12 +1,0 @@
-package de.thowl.notetilus.web;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import lombok.experimental.StandardException;
-
-@StandardException
-@ResponseStatus(value = HttpStatus.FORBIDDEN, reason = "Access Denied")
-public class ForbiddenException extends RuntimeException {
-
-}
