--- src/main/java/de/thowl/notetilus/web/NotesController.java
@@ -1,34 +1,0 @@
-package de.thowl.notetilus.web;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.SessionAttribute;
-
-import de.thowl.notetilus.core.services.AuthenticationService;
-import de.thowl.notetilus.storage.entities.AccessToken;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Controller
-public class NotesController {
-
-	@Autowired
-	private AuthenticationService authsvc;
-
-	@GetMapping("/u/{username}/notes")
-	public String showNotePage(@SessionAttribute(name = "token", required = false) AccessToken token,
-			@PathVariable("username") String username, Model model) {
-		log.info("entering showLoginPage (GET-Method: /login)");
-
-		// Prevent unauthrised access
-		if (!this.authsvc.validateSession(token, username))
-			throw new ForbiddenException("Unathorised access");
-
-		model.addAttribute("user", username);
-
-		return "notes";
-	}
-}
