--- src/main/java/de/thowl/notetilus/web/RegisterController.java
@@ -1,44 +1,0 @@
-package de.thowl.notetilus.web;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-
-import de.thowl.notetilus.core.services.AuthenticationService;
-import de.thowl.notetilus.web.forms.RegisterForm;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@Controller
-public class RegisterController {
-
-	@Autowired
-	private AuthenticationService authsvc;
-
-	@GetMapping("/register")
-	public String showRegisterPage() {
-		log.info("entering showRegisterPage (GET-Method: /register)");
-		return "register";
-	}
-
-	@PostMapping("/register")
-	public String doRegister(RegisterForm form, Model model) {
-		log.info("entering doRegister (POST-Method: /register)");
-
-		if (!authsvc.validateEmail(form.getEmail()))
-			model.addAttribute("error", "email_error");
-
-		if (!authsvc.validatePassword(form.getPassword()))
-			model.addAttribute("error", "password_error");
-
-		if (!form.getPassword().equals(form.getPassword2()))
-			model.addAttribute("error", "password_match_error");
-
-		this.authsvc.register(form.getFirstname(), form.getLastname(), form.getUsername(), form.getEmail(),
-				form.getPassword());
-
-		return "register";
-	}
-}
