--- src/main/java/de/thowl/notetilus/web/forms/LoginForm.java
@@ -1,10 +1,0 @@
-package de.thowl.notetilus.web.forms;
-
-import lombok.Data;
-
-@Data
-public class LoginForm {
-
-  private String email;
-  private String password;
-}
