+++ src/main/java/de/thowl/tnt/web/NotesController.java
@@ -1,0 +1,52 @@
+/*
+ *  TnT, Todo's 'n' Texts
+ *  Copyright (C) 2023  <name of author>
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package de.thowl.tnt.web;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.SessionAttribute;
+
+import de.thowl.tnt.core.services.AuthenticationService;
+import de.thowl.tnt.storage.entities.AccessToken;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Controller
+public class NotesController {
+
+	@Autowired
+	private AuthenticationService authsvc;
+
+	@GetMapping("/u/{username}/notes")
+	public String showNotePage(@SessionAttribute(name = "token", required = false) AccessToken token,
+			@PathVariable("username") String username, Model model) {
+		log.info("entering showLoginPage (GET-Method: /login)");
+
+		// Prevent unauthrised access
+		if (!this.authsvc.validateSession(token, username))
+			throw new ForbiddenException("Unathorised access");
+
+		model.addAttribute("user", username);
+
+		return "notes";
+	}
+}
