--- src/test/java/de/thowl/notetilus/NotetilusApplicationTests.java
@@ -1,15 +1,0 @@
-package de.thowl.notetilus;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class NotetilusApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
